move helpers etc to tools dir
