+++ src/py/idom/__init__.py
@@ -1,10 +1,7 @@
 __version__ = "0.1.2"
 
 from .core import element, Element, Events, Layout, SimpleServer, SharedServer
-from .bunch import StaticBunch, DynamicBunch
-from .helpers import Var, node, Image, hotswap
-from .display import display
-from . import html
+from .tools import StaticBunch, DynamicBunch, Var, node, Image, hotswap, display, html
 
 __all__ = [
     "element",
